Cleaned up the code
(cherry picked from commit 9ea3196)
