+++ src/main/java/sample/repository/MessageRepository.java
@@ -3,15 +3,8 @@ package sample.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import sample.model.Message;
-import sample.model.Room;
-import sample.model.User;
-
-import java.util.List;
 
 @Repository
 public interface MessageRepository extends JpaRepository<Message, Long> {
 
-    List<Message> findAllByRoom(Room room);
-
-    List<Message> findAllByUser(User user);
 }
