+++ src/main/java/sample/repository/UserRepository.java
@@ -6,5 +6,4 @@ import sample.model.User;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, String> {
-
 }
