+++ src/main/java/sample/service/MessageService.java
@@ -5,23 +5,16 @@ import sample.model.Message;
 import sample.model.Room;
 import sample.model.User;
 
-import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
 public interface MessageService {
 
-    List<Message> getAllMessages();
-
-    List<Message> getMessagesByUser(User user);
-
     Message createMessage(Message message);
 
-    List<Message> getAllMessagesInRoom(Room room);
-
-    void subscribeUser(Room room, User user, Map<User, WebSocketSession> sessionByUser);
-
     Map<Message, Set<User>> getReceivers();
+
+    Map<Room, Set<User>> getUsersByRoom();
 
     void sendMessageToRoom(Room room, Message message);
