+++ src/main/java/sample/service/SubscriptionService.java
@@ -1,0 +1,12 @@
+package sample.service;
+
+import org.springframework.web.socket.WebSocketSession;
+import sample.model.Room;
+import sample.model.User;
+
+import java.util.Map;
+
+public interface SubscriptionService {
+
+    void subscribeUser(Room room, User user, Map<User, WebSocketSession> sessionByUser);
+}
