+++ src/main/java/sample/service/impl/SubscriptionServiceImpl.java
@@ -2,77 +2,56 @@ package sample.service.impl;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.socket.TextMessage;
 import org.springframework.web.socket.WebSocketSession;
 import sample.model.Message;
 import sample.model.Room;
 import sample.model.User;
-import sample.model.UserRoomEntranceTime;
-import sample.repository.UserRoomEntranceTimeRepository;
-import sample.service.*;
+import sample.service.MessageService;
+import sample.service.SubscriptionService;
 
-import java.io.IOException;
 import java.time.LocalDateTime;
+import java.util.HashSet;
 import java.util.Map;
-import java.util.UUID;
+import java.util.Set;
 
 @Service
 public class SubscriptionServiceImpl implements SubscriptionService {
 
     private final MessageService messageService;
-    private final UserService userService;
-    private final RoomService roomService;
-    private final UserRoomEntranceTimeRepository entranceTimeRepository;
 
     @Autowired
-    public SubscriptionServiceImpl(MessageService messageService,
-                                   UserService userService,
-                                   RoomService roomService,
-                                   UserRoomEntranceTimeRepository entranceTimeRepository) {
+    public SubscriptionServiceImpl(MessageService messageService) {
         this.messageService = messageService;
-        this.userService = userService;
-        this.roomService = roomService;
-        this.entranceTimeRepository = entranceTimeRepository;
     }
 
     @Override
-    @Transactional(propagation = Propagation.REQUIRED)
-    public void subscribeUser(Room room, User user, Map<UUID, WebSocketSession> sessionByUser) {
+    @Transactional
+    public void subscribeUser(Room room, User user, Map<User, WebSocketSession> sessionByUser) {
 
-        if (!isUserSubscribed(room, user, sessionByUser)) {
-            roomService.setUserToRoom(user, room);
+        Map<Room, Set<User>> usersByRoom = messageService.getUsersByRoom();
 
-            String subscriptionMessage = "User " + user.getName() + " subscribed to the room " + room.getName();
-            Message message = MessageConstructor.constructMessage(user, room, subscriptionMessage, false);
+        Set<User> users = usersByRoom.computeIfAbsent(room, k -> new HashSet<>());
 
-            saveUserEntranceTime(room, user);
-
-            messageService.sendMessage(room, message, sessionByUser);
+        if (usersByRoom.get(room).contains(user)) {
+            return;
         }
+        users.add(user);
 
+        Message message = constructSubscribedMessage(user, room);
+
+        messageService.sendMessageToSubscribers(message, sessionByUser);
+        messageService.sendMessageToRoom(room, message);
     }
 
-    private void saveUserEntranceTime(Room room, User user) {
-        entranceTimeRepository.save(UserRoomEntranceTime.builder()
-                .roomName(room.getName())
-                .userUuid(user.getUuid())
-                .time(LocalDateTime.now())
-                .build());
+    private Message constructSubscribedMessage(User user, Room room) {
+        Message result = Message.builder()
+                .user(user)
+                .room(room)
+                .created(LocalDateTime.now())
+                .secret(false)
+                .text("User " + user.getName() + " subscribed to the room " + room.getName())
+                .build();
+        return messageService.createMessage(result);
     }
-
-    private boolean isUserSubscribed(Room room, User user, Map<UUID, WebSocketSession> sessionByUser) {
-        if (userService.containsUserInRoom(user, room)) {
-            try {
-                sessionByUser.get(user.getUuid()).sendMessage(new TextMessage("You already subscribed to room!"));
-                return true;
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        }
-        return false;
-    }
-
-
 }
